refactor: move and organize shared profile sync dependencies
