+++ packages/profile-sync-controller/src/controllers/user-storage/index.ts
@@ -4,5 +4,5 @@ const UserStorageController = Controller;
 export { Controller };
 export default UserStorageController;
 export * from './UserStorageController';
-export * from './encryption';
 export * as Mocks from './__fixtures__';
+export * from '../../shared/encryption';
