+++ packages/profile-sync-controller/src/sdk/authentication-jwt-bearer/types.ts
@@ -1,4 +1,4 @@
-import type { Env, Platform } from '../env';
+import type { Env, Platform } from '../../shared/env';
 
 export enum AuthType {
   /* sign in using a private key derived from your secret recovery phrase (SRP).
