--- packages/profile-sync-controller/src/sdk/encryption.ts
@@ -1,4 +1,0 @@
-import encryption from '../controllers/user-storage/encryption/encryption';
-
-export * from '../controllers/user-storage/encryption/encryption';
-export default encryption;
