+++ packages/profile-sync-controller/src/sdk/index.ts
@@ -1,6 +1,6 @@
 export * from './authentication';
 export * from './user-storage';
-export * from './env';
 export * from './errors';
-export * from './encryption';
 export * from './utils/messaging-signing-snap-requests';
+export * from '../shared/encryption';
+export * from '../shared/env';
