+++ packages/profile-sync-controller/src/shared/encryption/cache.ts
