+++ packages/profile-sync-controller/src/shared/encryption/encryption.test.ts
