+++ packages/profile-sync-controller/src/shared/encryption/encryption.ts
@@ -4,7 +4,7 @@ import { scryptAsync } from '@noble/hashes/scrypt';
 import { sha256 } from '@noble/hashes/sha256';
 import { utf8ToBytes, concatBytes, bytesToHex } from '@noble/hashes/utils';
 
-import type { NativeScrypt } from '../UserStorageController';
+import type { NativeScrypt } from '../types/encryption';
 import { getAnyCachedKey, getCachedKeyBySalt, setCachedKey } from './cache';
 import { base64ToByteArray, byteArrayToBase64, bytesToUtf8 } from './utils';
