+++ packages/profile-sync-controller/src/shared/encryption/index.ts
