+++ packages/profile-sync-controller/src/shared/encryption/utils.ts
