+++ packages/profile-sync-controller/src/shared/storage-schema.ts
