+++ packages/profile-sync-controller/src/shared/types/encryption.ts
@@ -1,0 +1,8 @@
+export declare type NativeScrypt = (
+  passwd: string,
+  salt: Uint8Array,
+  N: number,
+  r: number,
+  p: number,
+  size: number,
+) => Promise<Uint8Array>;
